Separate the rendering process from the canvas. Abstract away rendering
The rendering now has no concept of the underlying Canvas object. It is hidden behind the Window. Rendering is done separately to avoid calling things that should not be called. This makes it more easy to define renderers without touching any "native" AWT or Swing code.

As a result, game-agnostic rendering concepts have been moved to an engine package.